Fixing wrong sprintf() calls.
modified:
  storage/connect/filamap.cpp
  storage/connect/filamdbf.cpp
  storage/connect/filamfix.cpp
  storage/connect/filamtxt.cpp
  storage/connect/filamvct.cpp
  storage/connect/ha_connect.cc
  storage/connect/osutil.c
  storage/connect/plgdbutl.cpp
  storage/connect/plugutil.c
  storage/connect/tabfix.cpp
  storage/connect/tabxml.cpp
  storage/connect/user_connect.cc
  storage/connect/value.cpp
  storage/connect/xindex.cpp

+++ storage/connect/filamap.cpp
@@ -164,7 +164,7 @@ bool MAPFAM::OpenTableFile(PGLOBAL g)
 
       if (!(*g->Message))
         sprintf(g->Message, MSG(OPEN_MODE_ERROR),
-                "map", rc, filename);
+                "map", (int) rc, filename);
 
 #ifdef DEBTRACE
  htrc("%s\n", g->Message);
